requirements.txt
docx==0.2.4
watson_developer_cloud==1.0.2
scipy==0.19.1
Pillow==5.0.0
